addressed comments, moved the functions to consensus/bor
